+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - **[FIX]** `SuccessModal` display with small content
+- **[UPDATE]** Allow `format` prop on TextField component
 
 # v0.20.3 (09/01/2019)